refactor(components): moved field.base from service to components
field.base as classes which will be foundation for components thus is moved from service to

components
