+++ src/main.ts
@@ -1,4 +1,4 @@
-import {FieldBase} from "./services/field.base";
+import {FieldBase} from "./components/field.base";
 import {FormlyConfig} from "./services/formly.config";
 import {FormlyPubSub, FormlyEventEmitter} from "./services/formly.event.emitter";
 import {ControlService} from "./services/control.service";
@@ -8,7 +8,7 @@ import {FormlyForm} from "./components/formly.form";
 import {FormlyField} from "./components/formly.field";
 import {FormlyProviders} from "./services/formly.providers";
 
-export {FieldBase} from "./services/field.base";
+export {FieldBase} from "./components/field.base";
 export {FormlyConfig} from "./services/formly.config";
 export {FormlyPubSub, FormlyEventEmitter} from "./services/formly.event.emitter";
 export {ControlService} from "./services/control.service";
