+++ src/services/control.service.ts
@@ -1,6 +1,6 @@
 import {Injectable}   from "@angular/core";
 import {FormBuilder, ControlGroup, RadioButtonState} from "@angular/common";
-import {FieldBase} from "./field.base";
+import {FieldBase} from "../components/field.base";
 
 @Injectable()
 export class ControlService {
